Move WebSocketConnectionError to exceptions
