+++ ariadne/types.py
@@ -47,7 +47,6 @@ __all__ = [
     "OnDisconnect",
     "OnOperation",
     "OnComplete",
-    "WebSocketConnectionError",
     "Extension",
     "SchemaBindable",
 ]
@@ -501,18 +500,6 @@ Called with two arguments:
 `Operation`: an object with closed subscription's data.
 """
 OnComplete = Callable[[WebSocket, Operation], Any]
-
-
-class WebSocketConnectionError(Exception):
-    """Special error class enabling custom error reporting for on_connect"""
-
-    def __init__(self, payload: Optional[Union[dict, str]] = None) -> None:
-        if isinstance(payload, dict):
-            self.payload = payload
-        elif payload:
-            self.payload = {"message": str(payload)}
-        else:
-            self.payload = {"message": "Unexpected error has occurred."}
 
 
 class Extension:
